Move run_* from main to lib in order to support tests.
