+++ src/lib.rs
@@ -3,6 +3,7 @@ extern crate pest_derive;
 
 pub mod app_cli;
 pub mod app_config;
+pub mod app_run;
 pub mod help;
 pub mod parser;
 pub mod sic_processor;
